Add useOfferForm
