+++ src/components/Offers/New/form-components/OfferForm.js
@@ -1,0 +1,365 @@
+import {
+    CardContent,
+    CardHeader,
+    DialogContent,
+    FormHelperText,
+    Grid,
+    FormControl,
+    Typography,
+    Collapse,
+    Button,
+} from "@material-ui/core";
+import React, { useState, useCallback, useContext } from "react";
+import { Redirect } from "react-router-dom";
+import MultiOptionTextField from "../../../utils/form/MultiOptionTextField";
+import MultiOptionAutocomplete from "../../../utils/form/MultiOptionAutocomplete";
+import ConnectedLoginAlert from "../../../utils/LoginAlert";
+import TitleComponent from "./TitleComponent";
+import OwnerComponent from "./OwnerComponent";
+import LocationComponent from "./LocationComponent";
+import JobTypeComponent from "./JobTypeComponent";
+import JobStartDateComponent from "./JobStartDateComponent";
+import JobDurationComponent from "./JobDurationComponent";
+import VacanciesComponent from "./VacanciesComponent";
+import IsPaidComponent from "./IsPaidComponent";
+import PublicationDateComponent from "./PublicationDateComponent";
+import PublicationEndDateComponent from "./PublicationEndDateComponent";
+import IsHiddenComponent from "./IsHiddenComponent";
+import TextEditorComponent from "./TextEditorComponent";
+import {
+    KeyboardArrowDown,
+    KeyboardArrowUp,
+} from "@material-ui/icons";
+import useOfferFormStyles from "./offerStyles";
+import { Controller } from "react-hook-form";
+import { useMobile } from "../../../../utils/media-queries";
+import "../editor.css";
+
+export const PAID_OPTIONS = [
+    { value: "none", label: "Unspecified" },
+    { value: true, label: "Paid" },
+    { value: false, label: "Unpaid" },
+];
+
+const OfferForm = (Context) => {
+    const {
+        submit,
+        errors,
+        requestErrors,
+        control,
+        fields,
+        contacts,
+        requirements,
+        getValues,
+        appendContact,
+        removeContact,
+        appendRequirement,
+        removeRequirement,
+        fieldsSelectorProps,
+        techSelectorProps,
+        loading,
+        success,
+        offerId,
+        isAdmin,
+        isLoggedIn,
+        companyUnfinishedRegistration,
+        showCompanyField,
+        showHiddenField,
+    } = useContext(Context);
+
+    const isMobile = useMobile();
+
+    const disabled = !isLoggedIn || companyUnfinishedRegistration;
+
+    const [isAdvancedOpen, setAdvancedOpen] = useState(false);
+
+    const isAdvancedOpenOrErrors = useCallback(() => (
+        isAdvancedOpen ||
+        !!errors.publishDate ||
+        !!errors.publishEndDate ||
+        !!errors.isHidden
+    ), [errors.isHidden, errors.publishDate, errors.publishEndDate, isAdvancedOpen]);
+
+    const Content = isMobile ? DialogContent : CardContent;
+    const classes = useOfferFormStyles(isMobile)();
+
+    const showOwnerComponent = isAdmin && showCompanyField;
+
+    const SelectStylingProps = {
+        SelectProps: {
+            classes: {
+                select: classes.menuSelect,
+            },
+        },
+    };
+
+    return (
+        success
+            ? <Redirect to={`/offer/${offerId}`} push />
+            :
+            <div className={classes.formCard}>
+                <CardHeader title={!isMobile && "New Offer" } />
+                <Content className={classes.formContent}>
+                    <ConnectedLoginAlert
+                        isLoggedIn={isLoggedIn}
+                        companyUnfinishedRegistration={companyUnfinishedRegistration}
+                    />
+                    <Grid container className={classes.formArea}>
+                        <Grid item xs={12}>
+                            <form
+                                onSubmit={submit}
+                                aria-label="Create Offer Form"
+                            >
+                                <Grid container spacing={4}>
+                                    <Grid item xs={12} lg={isAdmin ? 6 : 12}>
+                                        <TitleComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
+                                            control={control}
+                                        />
+                                    </Grid>
+                                    {isAdmin && showOwnerComponent &&
+                                        <Grid item xs={12} lg={6}>
+                                            <OwnerComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
+                                                control={control}
+                                            />
+                                        </Grid>}
+                                    <Grid item xs={12} lg={6}>
+                                        <FormControl fullWidth margin="dense">
+                                            <LocationComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
+                                                control={control}
+                                            />
+                                        </FormControl>
+                                    </Grid>
+                                    <Grid item xs={12} lg={6} className={classes.jobTypeGrid}>
+                                        <FormControl fullWidth margin="dense">
+                                            <JobTypeComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
+                                                control={control}
+                                                textFieldProps={{
+                                                    ...SelectStylingProps,
+                                                }}
+                                            />
+                                        </FormControl>
+                                    </Grid>
+                                    <Grid item xs={12} lg={6}>
+                                        <Controller
+                                            name="fields"
+                                            render={(
+                                                { field: {  onBlur, name } },
+                                            ) => (
+                                                <MultiOptionAutocomplete
+                                                    name={name}
+                                                    onBlur={onBlur}
+                                                    error={errors.fields || requestErrors.fields}
+                                                    disabled={disabled}
+                                                    chipWrapperProps={{
+                                                        className: classes.autocompleteChipWrapper,
+                                                    }}
+                                                    textFieldProps={{
+                                                        margin: "none",
+                                                    }}
+                                                    {...fieldsSelectorProps}
+                                                    label="Fields *"
+                                                    placeholder="Fields *"
+                                                />
+                                            )}
+                                            control={control}
+                                        />
+                                    </Grid>
+                                    <Grid item xs={12} lg={6}>
+                                        <Controller
+                                            name="technologies"
+                                            render={(
+                                                { field: { onBlur, name } },
+                                            ) => (
+                                                <MultiOptionAutocomplete
+                                                    name={name}
+                                                    onBlur={onBlur}
+                                                    error={errors.technologies || requestErrors.technologies}
+                                                    disabled={disabled}
+                                                    chipWrapperProps={{
+                                                        className: classes.autocompleteChipWrapper,
+                                                    }}
+                                                    textFieldProps={{
+                                                        margin: "none",
+                                                    }}
+                                                    {...techSelectorProps}
+                                                    label="Technologies *"
+                                                    placeholder="Technologies *"
+                                                />)}
+                                            control={control}
+                                        />
+                                    </Grid>
+                                    <Grid item xs={12} lg={6}>
+                                        <FormControl fullWidth>
+                                            <JobStartDateComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
+                                                control={control}
+                                            />
+                                        </FormControl>
+                                    </Grid>
+                                    <Grid item xs={12} lg={6}>
+                                        <JobDurationComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
+                                            control={control}
+                                        />
+                                    </Grid>
+                                    <Grid item xs={12} lg={6} className={classes.vacanciesGrid}>
+                                        <FormControl fullWidth>
+                                            <VacanciesComponent
+                                                disabled={disabled}
+                                                errors={errors}
+                                                requestErrors={requestErrors}
+                                                control={control}
+                                            />
+                                        </FormControl>
+                                    </Grid>
+                                    <Grid item xs={12} lg={6}>
+                                        <IsPaidComponent
+                                            disabled={disabled}
+                                            errors={errors}
+                                            requestErrors={requestErrors}
+                                            control={control}
+                                            textFieldProps={{
+                                                ...SelectStylingProps,
+                                            }}
+                                        />
+                                    </Grid>
+                                    <Grid item xs={12} lg={12}>
+                                        <Button
+                                            onClick={() => setAdvancedOpen(!isAdvancedOpen)}
+                                            size="small"
+                                            margin="dense"
+                                            endIcon={
+                                                isAdvancedOpenOrErrors()
+                                                    ? <KeyboardArrowUp />
+                                                    : <KeyboardArrowDown />}
+                                        >
+                                            <Typography>Advanced Settings</Typography>
+                                        </Button>
+                                    </Grid>
+                                    <Grid item xs={12} lg={12}>
+                                        <Collapse in={isAdvancedOpenOrErrors()}>
+                                            <Grid container spacing={4} className={classes.advancedSettingsCollapse}>
+                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
+                                                    <PublicationDateComponent
+                                                        disabled={disabled}
+                                                        errors={errors}
+                                                        requestErrors={requestErrors}
+                                                        control={control}
+                                                        datePickerProps={{
+                                                            className: classes.advancedSettingsDatePicker,
+                                                        }}
+                                                    />
+                                                </Grid>
+                                                <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
+                                                    <PublicationEndDateComponent
+                                                        fields={fields}
+                                                        disabled={disabled}
+                                                        errors={errors}
+                                                        requestErrors={requestErrors}
+                                                        control={control}
+                                                        datePickerProps={{
+                                                            className: classes.advancedSettingsDatePicker,
+                                                        }}
+                                                    />
+                                                </Grid>
+                                                {
+                                                    showHiddenField &&
+                                                    <Grid item xs={12} lg={6} className={classes.gridWithInfo}>
+                                                        <IsHiddenComponent
+                                                            disabled={disabled}
+                                                            errors={errors}
+                                                            requestErrors={requestErrors}
+                                                            control={control}
+                                                        />
+                                                    </Grid>
+                                                }
+                                            </Grid>
+                                        </Collapse>
+                                    </Grid>
+                                    <Grid item xs={12} className={classes.multiTextOptionGrid}>
+                                        <MultiOptionTextField
+                                            values={contacts}
+                                            label="Contacts *"
+                                            itemLabelPrefix="Contact #"
+                                            controllerName="contacts"
+                                            onAdd={appendContact}
+                                            onRemove={removeContact}
+                                            getValues={getValues}
+                                            control={control}
+                                            errors={errors.contacts || requestErrors.contacts}
+                                            disabled={disabled}
+                                            addEntryBtnTestId="contacts-selector"
+                                        />
+                                    </Grid>
+                                    <Grid item xs={12} className={classes.multiTextOptionGrid}>
+                                        <MultiOptionTextField
+                                            values={requirements}
+                                            label="Requirements *"
+                                            itemLabelPrefix="Requirement #"
+                                            controllerName="requirements"
+                                            onAdd={appendRequirement}
+                                            onRemove={removeRequirement}
+                                            getValues={getValues}
+                                            control={control}
+                                            errors={errors.requirements || requestErrors.requirements}
+                                            disabled={disabled}
+                                            textFieldProps={{ multiline: true }}
+                                            addEntryBtnTestId="requirements-selector"
+                                        />
+                                    </Grid>
+                                </Grid>
+                                <TextEditorComponent
+                                    fields={fields}
+                                    disabled={disabled}
+                                    errors={errors}
+                                    requestErrors={requestErrors}
+                                    control={control}
+                                />
+                                {requestErrors.generalErrors ?
+                                    requestErrors.generalErrors.map((error, idx) => (
+                                        <FormHelperText key={`${error.message}-${idx}`} error>
+                                            {error.message}
+                                        </FormHelperText>
+                                    ))
+                                    :
+                                    <FormHelperText error={true}>
+                                        {" "}
+                                    </FormHelperText>
+                                }
+                                <Button
+                                    disabled={loading || disabled}
+                                    variant="contained"
+                                    color="primary"
+                                    onClick={submit}
+                                    data-testid="submit-offer"
+                                >
+                                        Submit
+                                </Button>
+                                <div className={classes.requiredFields}>
+                                    <p>* Required fields</p>
+                                </div>
+                            </form>
+                        </Grid>
+                    </Grid>
+                </Content>
+            </div>
+    );
+};
+
+export default OfferForm;
