+++ src/pages/EditOfferPage.js
@@ -1,0 +1,10 @@
+import React from "react";
+import EditOfferForm from "../components/Offers/EditOfferForm";
+
+const EditOfferPage = () => (
+    <>
+        <EditOfferForm />
+    </>
+);
+
+export default EditOfferPage;
