+++ src/services/offerService.js
@@ -240,3 +240,65 @@ export const newOffer = async ({
         throw [{ msg: "Unexpected Error. Please try again later." }];
     }
 };
+
+export const editOffer = async ({
+    offerId,
+    title,
+    publishDate,
+    publishEndDate,
+    jobMinDuration,
+    jobMaxDuration,
+    jobStartDate,
+    description,
+    contacts,
+    isPaid,
+    vacancies,
+    jobType,
+    fields,
+    technologies,
+    location,
+    coordinates,
+    requirements,
+}) => {
+    const data = {
+        title,
+        publishDate,
+        publishEndDate,
+        jobMinDuration,
+        jobMaxDuration,
+        jobStartDate,
+        description,
+        contacts,
+        isPaid,
+        vacancies,
+        jobType,
+        fields,
+        technologies,
+        location,
+        coordinates,
+        requirements,
+    };
+
+    try {
+        const res = await fetch(`${API_HOSTNAME}/offers/edit/${offerId}`, {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            credentials: "include",
+            body: JSON.stringify(data),
+        });
+        const json = await res.json();
+
+        if (!res.ok) {
+            throw json.errors;
+        }
+        // TODO count metrics
+        return json;
+
+    } catch (error) {
+        // TODO count metrics
+        if (Array.isArray(error)) throw error;
+        throw [{ msg: "Unexpected Error. Please try again later." }];
+    }
+};
